# docs/kaayko-paddling-api-swagger.yaml
openapi: 3.0.3
info:
  title: Kaayko Paddling Intelligence APIs
  version: "2.1.0"
  description: |
    # Kaayko Paddling Intelligence APIs v2.1.0
    
    **Enterprise-grade paddling condition intelligence powered by machine learning and real-time marine data integration**
    
    ---
    
    ## 🎯 Executive Summary
    
    The Kaayko Paddling Intelligence APIs provide ultra-fast, machine learning-powered paddling condition assessments with comprehensive safety analysis. Our platform processes real-time weather, marine, and environmental data to deliver actionable insights for paddlers across 18+ premium locations.
    
    ### Core Capabilities
    
    | API Endpoint | Purpose | Response Time | Key Features |
    |--------------|---------|---------------|--------------|
    | `paddlingOut` | Location Discovery | ~100ms | Comprehensive location database with amenities |
    | `paddleScore` | Real-time Conditions | ~250ms | Current conditions with ML-powered safety ratings |
    | `fastForecast` | 3-Day Predictions | <4s fresh / <1s cached | Hourly forecasts with penalty breakdown |
    | `forecast` | Premium Analytics | Variable | Advanced analytics for professional use |
    
    ### 🚀 Version 2.1.0 Enhancements
    
    - **🎯 Precision Rating System**: 0.5-increment ratings (1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0)
    - **🌊 Marine Data Integration**: Real-time wave height and water temperature analysis
...
    
    ## 🧠 Machine Learning Intelligence
    
...
    ```
...
    
...
servers:
...
security:
  - ApiKeyAuth: []
...
paths:
...
  /fastForecast:
    get:
      tags:
        - fastForecast
      operationId: getFastForecast
      summary: Get 3-day weather forecast with ML-powered paddle predictions
      description: |
        **Ultra-fast 72-hour forecast with machine learning paddle intelligence and marine data integration**
        
        This high-performance endpoint delivers comprehensive 3-day hourly forecasts with ML-generated paddle ratings, safety warnings, and detailed penalty analysis.
        
        ### 🚀 Performance Characteristics
        - **Fresh Data**: <4 seconds (67% faster than v2.0)
        - **Cached Data**: <1 second (sub-second for repeated requests)
        - **Production ML**: Powered by kaayko-ml-service-87383373015.us-central1.run.app
        
        ### 🧠 Advanced ML Features
        - **72-Hour Predictions**: Hourly paddle ratings for full 3 days
        - **Marine Integration**: Wave height and water temperature when available
        - **Penalty Transparency**: Detailed breakdown of all rating adjustments
        - **Safety Warnings**: Proactive alerts for dangerous conditions
        - **Beaufort Scale**: Professional wind analysis (0-12 scale)
        - **UV Analysis**: Comprehensive sun exposure assessment
        
        ### 📊 Enhanced Penalty System
        ```yaml
        Temperature Penalties:
          High Heat (85-94°F):    -1.0 rating points
          Extreme Heat (95°F+):   -1.5 rating points
          
        UV Index Penalties:
          Moderate (5-8):         -0.5 rating points
          High (9-10):           -0.5 rating points  
          Dangerous (11+):       -1.0 rating points
          
        Wind Penalties:
          Strong Winds (25+ mph): -2.0 rating points
        ```
        
        ### 🔄 Cross-API Consistency
        Identical weather conditions will produce identical paddle ratings across `paddleScore` and `fastForecast` endpoints, ensuring reliable decision-making.
        
        ### ⚠️ Marine Data Requirement
        **Always use coordinates from `paddlingOut` endpoint** to ensure marine data availability and prediction accuracy.
        
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
            format: double
            minimum: -90.0
            maximum: 90.0
            example: 32.881187
          description: |
            **Latitude coordinate in decimal degrees**
            
            Must be a valid latitude between -90.0 and 90.0. 
            
            **Example**: `32.881187` (Trinity River, Texas)
        - name: lng
          in: query
          required: true
          schema:
            type: number
            format: double
            minimum: -180.0
            maximum: 180.0
            example: -96.929937
          description: |
            **Longitude coordinate in decimal degrees**
            
            Must be a valid longitude between -180.0 and 180.0.
            
            **Example**: `-96.929937` (Trinity River, Texas)
      responses:
        '200':
          description: Successfully retrieved 3-day forecast with ML predictions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FastForecastResponse'
              examples:
                perfect_mountain_conditions:
                  summary: Perfect Mountain Conditions - Colorado
                  description: High altitude location with excellent paddling conditions and no penalties
                  value:
                    success: true
                    location:
                      name: "Buena Vista"
                      region: "Colorado"
                      country: "United States of America"
                      coordinates:
                        latitude: 38.781063
                        longitude: -106.277812
                    forecast:
                      - date: "2025-08-23"
                        hourly:
                          8:
                            temperature: 65.8
                            windSpeed: 4.2
                            windDirection: "SW"
                            gustSpeed: 5.5
                            humidity: 35
                            cloudCover: 10
                            uvIndex: 3.2
                            visibility: 25
                            hasWarnings: false
                            warnings: []
                            beaufortScale: 2
                            waveHeight: null
                            waterTemp: null
                            marineDataAvailable: false
                            prediction:
                              rating: 5.0
                              originalRating: 5.0
                              penaltiesApplied: []
                              totalPenalty: 0.0
                              roundedTo05Increments: true
                              mlModelUsed: true
                              predictionSource: "ml-model"
                            rating: 5.0
                            mlModelUsed: true
                            predictionSource: "ml-model"
                          15:
                            temperature: 78.2
                            windSpeed: 8.1
                            windDirection: "W"
                            gustSpeed: 10.5
                            humidity: 28
                            cloudCover: 5
                            uvIndex: 8.1
                            visibility: 25
                            hasWarnings: false
                            warnings: []
                            beaufortScale: 3
                            waveHeight: null
                            waterTemp: null
                            marineDataAvailable: false
                            prediction:
                              rating: 4.5
                              originalRating: 4.5
                              penaltiesApplied: []
                              totalPenalty: 0.0
                              roundedTo05Increments: true
                              mlModelUsed: true
                              predictionSource: "ml-model"
                            rating: 4.5
                            mlModelUsed: true
                            predictionSource: "ml-model"
                    metadata:
                      processingTimeMs: 3247
                      responseTime: "3.25s"
                      cached: false
                      mlServiceUrl: "https://kaayko-ml-service-87383373015.us-central1.run.app"
                      source: "coordinate_cache"
                      fastAPI: true
                      timestamp: "2025-08-23T15:45:22.120Z"
                
                extreme_heat_warnings:
                  summary: Extreme Heat with Multiple Penalties - Texas Summer
                  description: Shows multiple penalty system in action with detailed breakdown
                  value:
                    success: true
                    location:
                      name: "Irving"
                      region: "Texas"
                      country: "United States of America"
                      coordinates:
                        latitude: 32.881187
                        longitude: -96.929937
                    forecast:
                      - date: "2025-08-23"
                        hourly:
                          15:
                            temperature: 99.2
                            windSpeed: 1.8
                            windDirection: "S"
                            gustSpeed: 2.3
                            humidity: 38
                            cloudCover: 15
                            uvIndex: 9.8
                            visibility: 18
                            hasWarnings: true
                            warnings: [
                              "DANGER: Extreme heat (99°F) - high risk of heat exhaustion",
                              "DANGER: Very high UV (9.8) - severe burn risk"
                            ]
                            beaufortScale: 1
                            waveHeight: 0.1
                            waterTemp: 29.8
                            marineDataAvailable: true
                            prediction:
                              rating: 2.0
                              originalRating: 4.0
                              penaltiesApplied: [
                                "Extreme heat (99.2°F): -1.5",
                                "High UV (9.8): -0.5"
                              ]
                              totalPenalty: 2.0
                              roundedTo05Increments: true
                              mlModelUsed: true
                              predictionSource: "ml-model"
                            rating: 2.0
                            mlModelUsed: true
                            predictionSource: "ml-model"
                    metadata:
                      processingTimeMs: 3891
                      responseTime: "3.89s"
                      cached: false
                      mlServiceUrl: "https://kaayko-ml-service-87383373015.us-central1.run.app"
                      source: "api_fresh"
                      fastAPI: true
                      timestamp: "2025-08-23T15:45:22.120Z"
                
                cached_response:
                  summary: Cached Response - Sub-second Performance
                  description: Example of cached response showing sub-second performance
                  value:
                    success: true
                    location:
                      name: "Port Angeles"
                      region: "Washington"
                      country: "United States of America"
                      coordinates:
                        latitude: 48.052813
                        longitude: -123.870438
                    forecast:
                      - date: "2025-08-23"
                        hourly:
                          12:
                            temperature: 70.1
                            windSpeed: 8.2
                            beaufortScale: 3
                            uvIndex: 5.1
                            prediction:
                              rating: 4.5
                              originalRating: 4.5
                              penaltiesApplied: []
                              totalPenalty: 0.0
                              roundedTo05Increments: true
                              mlModelUsed: true
                              predictionSource: "ml-model"
                            rating: 4.5
                            mlModelUsed: true
                            predictionSource: "ml-model"
                    metadata:
                      processingTimeMs: 847
                      responseTime: "0.85s"
                      cached: true
                      source: "coordinate_cache"
                      fastAPI: true
                      timestamp: "2025-08-23T15:45:22.120Z"
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_location:
                  summary: Missing location parameters
                  value:
                    success: false
                    error: "Invalid input parameters"
                    details: ["Location required: provide coordinates (lat,lng) or spotId"]
                    usage:
                      endpoint: "/fastForecast"
                      parameters:
                        location_methods: [
                          "lat=32.881187&lng=-96.929937 (separate coordinates)",
                          "location=32.881187,-96.929937 (combined coordinates)",
                          "spotId=trinity_river (known spot ID)"
                        ]
                      examples:
                        coordinates: "/fastForecast?lat=32.881187&lng=-96.929937"
                        location_string: "/fastForecast?location=32.881187,-96.929937"
                        spot_id: "/fastForecast?spotId=trinity_river"
                invalid_latitude:
                  summary: Invalid latitude parameter
                  value:
                    success: false
                    error: "Invalid input parameters"
                    details: ["Latitude must be between -90 and 90"]
                    usage:
                      endpoint: "/fastForecast"
                      examples:
                        valid_coordinates: "/fastForecast?lat=32.881187&lng=-96.929937"
        '503':
          description: Service temporarily unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Forecast service unavailable"
                details: "Weather data source temporarily unavailable"
                suggestion: "Try again in a few minutes"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        **Enterprise API Key Authentication**
        
        Contact api-support@kaayko.com for enterprise access and API key provisioning.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        **OAuth 2.0 Bearer Token Authentication**
        
        For service-to-service authentication and enterprise integrations.
...
  schemas:
...
    FastForecastResponse:
      type: object
      title: Fast Forecast Response
      description: High-performance 3-day forecast with ML paddle predictions and marine data
      required:
        - success
        - location
        - forecast
        - metadata
      properties:
        success:
          type: boolean
          description: Request success indicator
          example: true
        location:
          type: object
          title: Location Information
          description: Detailed location context and geographical information
          required:
            - name
            - region
            - country
            - coordinates
          properties:
            name:
              type: string
              description: City or locality name
              example: "Irving"
            region:
              type: string
              description: State or province
              example: "Texas"
            country:
              type: string
              description: Country name
              example: "United States of America"
            coordinates:
              type: object
              description: Exact coordinates used for forecast
              required:
                - latitude
                - longitude
              properties:
                latitude:
                  type: number
                  format: double
                  description: Latitude in decimal degrees
                  example: 32.881187
                longitude:
                  type: number
                  format: double
                  description: Longitude in decimal degrees
                  example: -96.929937
            elevation:
              type: integer
              description: Elevation above sea level in feet
              example: 450
            timezone:
              type: string
              description: IANA timezone identifier
              example: "America/Chicago"
        forecast:
          type: array
          description: 3 days of comprehensive hourly forecasts with ML predictions
          minItems: 1
          maxItems: 3
          items:
            type: object
            title: Daily Forecast
            description: Complete daily forecast with hourly breakdown and summary
            required:
              - date
              - hourly
            properties:
              date:
                type: string
                format: date
                description: Forecast date in YYYY-MM-DD format
                example: "2025-08-23"
              summary:
                type: object
                title: Daily Summary
                description: Daily aggregated forecast metrics and insights
                properties:
                  maxTemp:
                    type: number
                    format: float
                    description: Maximum temperature for the day
                    example: 95.2
                  minTemp:
                    type: number
                    format: float
                    description: Minimum temperature for the day
                    example: 78.8
                  avgWindSpeed:
                    type: number
                    format: float
                    description: Average wind speed for the day
                    example: 6.8
                  precipitation:
                    type: number
                    format: float
                    description: Total precipitation in inches
                    example: 0.0
                  paddleWindow:
                    type: string
                    description: Recommended paddling time window
                    example: "Early morning (6-9 AM) and evening (5-7 PM)"
                  bestRating:
                    type: number
                    format: float
                    multipleOf: 0.5
                    minimum: 1.0
                    maximum: 5.0
                    description: Highest paddle rating for the day
                    example: 4.0
                  avgRating:
                    type: number
                    format: float
                    description: Average paddle rating for the day
                    example: 3.2
                  conditions:
                    type: string
                    description: General weather conditions summary
                    example: "Hot and sunny with light winds"
              hourly:
                type: object
                description: Hourly forecast data keyed by hour (00-23)
                additionalProperties:
                  $ref: '#/components/schemas/HourlyForecast'
                example:
                  "08":
                    temperature: 82.4
                    windSpeed: 4.2
                    rating: 4.0
                  "15":
                    temperature: 99.2
                    windSpeed: 1.8
                    rating: 2.0
        metadata:
          type: object
          title: Response Metadata
          description: Processing information and API performance metrics
          required:
            - processingTimeMs
            - responseTime
            - cached
          properties:
            processingTimeMs:
              type: integer
              description: Total processing time in milliseconds
              minimum: 0
              example: 3833
            responseTime:
              type: string
              description: Human-readable response time
              pattern: '^\d+(\.\d+)?(ms|s)$'
              example: "3.83s"
            cached:
              type: boolean
              description: Whether response was served from cache
              example: false
            cacheAge:
              type: string
              description: Age of cached data (if applicable)
              example: "14 minutes"
            mlServiceUrl:
              type: string
              format: uri
              description: Production ML service endpoint used
              example: "https://kaayko-ml-service-87383373015.us-central1.run.app"
            dataFreshness:
              type: string
              description: Data freshness indicator
              enum: ["live", "cached", "stale"]
              example: "live"
            totalHours:
              type: integer
              description: Total hours of forecast data provided
              example: 72
            apiVersion:
              type: string
              description: API version used for processing
              example: "2.1.0"
...
    ErrorResponse:
      type: object
      title: Error Response
      description: Standardized error response format with detailed debugging information
      required:
        - success
        - error
        - code
        - timestamp
      properties:
        success:
          type: boolean
          description: Request success indicator (always false for errors)
          example: false
        error:
          type: string
          description: Human-readable error message
          maxLength: 200
          example: "Invalid coordinates"
        code:
          type: string
          description: Machine-readable error code for programmatic handling
          pattern: '^[A-Z_]+$'
          example: "INVALID_COORDINATES"
        details:
          type: string
          description: Additional technical details about the error
          maxLength: 500
          example: "Latitude must be between -90.0 and 90.0, longitude must be between -180.0 and 180.0"
        suggestion:
          type: string
          description: Helpful suggestion for resolving the error
          maxLength: 300
          example: "Use coordinates from paddlingOut API for accurate marine data"
        timestamp:
          type: string
          format: date-time
          description: UTC timestamp when error occurred
          example: "2025-08-23T15:45:22.120Z"
        requestId:
          type: string
          description: Unique identifier for this request (for support purposes)
          pattern: '^[a-f0-9-]+$'
          example: "req_8f9e2d1a-4b5c-6789-01ef-234567890abc"
        retryAfter:
          type: integer
          minimum: 0
          description: Seconds to wait before retrying (for rate limit errors)
          example: 60
        supportUrl:
          type: string
          format: uri
          description: URL for additional support or documentation
          example: "https://kaayko.com/docs/api/errors#invalid-coordinates"
...